final anggota controller

+++ app/Http/Controllers/Admin/Team/AnggotaController.php
@@ -11,63 +11,62 @@ use Illuminate\Support\Facades\Storage;
 
 class AnggotaController extends Controller
 {
-    /* FUNGSI VIEW ANGGOTA */
     public function index()
     {
         return view('admin.team.index',[
-            'title' => 'Teams',
-            'teams' => Anggota::with('jabatans')->get(),
-            'jabatans' => KategoriJabatan::all()
+            'title'     => 'Teams',
+            'teams'     => Anggota::with('jabatans')->get(),
+            'jabatans'  => KategoriJabatan::all()
         ]);
     }
 
-    /* FUNGSI TAMBAH ANGGOTA */
     public function store(AnggotaRequest $request)
     {
         if ($request->ajax()) {
-            # code...
+
             if ($request->has('foto_anggota')) {
+
                 $imgName = date('HisdmY').'_'.str_replace(' ','_',strtolower($request->nama_anggota)).'.'.$request->foto_anggota->extension();
                 $pathName = Storage::putFileAs('public/teams', $request->file('foto_anggota'), $imgName);
-                // try {
+                
+                try {
+                    Anggota::create([
+                        ...$request->all(),
+                        'jabatan_id'    => (int) $request->jabatan_id,
+                        'foto_anggota'  => $pathName,
+                    ]);
+                    return $this->successResponse(null,'created');
+
+                } catch (\Throwable $th) {
+                    throw $th;
+                }
+
+            }else{
+                try {
                     Anggota::create([
                         ...$request->all(),
                         'jabatan_id' => (int) $request->jabatan_id,
-                        'foto_anggota' => $pathName,
                     ]);
-    
-                    return response()->json([
-                        'message' => 'Data Created Successfully',
-                        'data' => $request->all(),
-                    ]);
-    
-                // } catch (\Throwable $th) {
-                //     //throw $th;
-                // }
-        }else{
-            Anggota::create([
-                ...$request->all(),
-                'jabatan_id' => (int) $request->jabatan_id,
-            ]);
-
-            return response()->json([
-                'message' => 'Data Created Successfully',
-                'data' => $request->all(),
-            ]);
-        }
-            
+                    return $this->successResponse(null,'created');
+        
+                } catch (\Throwable $th) {
+                    throw $th;
+                }
+            }
         }
     }
 
-    /* FUNGSI LIHAT DATA ANGGOTA */
+    /* FUNGSI LIHAT DATA ANGGOTA BY ID */
     public function show($anggota)
     {
-        return response()->json([
-            'data' => Anggota::findOrfail($anggota)
-        ]);
+        if (request()->ajax()) {
+            return response()->json([
+                'data' => Anggota::findOrfail($anggota)
+            ]);
+        }
+        abort(404);
     }
 
-    /* FUNGSI UPDATE ANGGOTA */
     public function update(Request $request, $anggota)
     {
         if ($request->ajax()) {
@@ -84,40 +83,35 @@ class AnggotaController extends Controller
                 
                 $update = $data->update([
                     ...$request->all(),
-                    'jabatan_id' => (int) $request->jabatan_id,
-                    'foto_anggota' => $pathName,
+                    'jabatan_id'    => (int) $request->jabatan_id,
+                    'foto_anggota'  => $pathName,
                 ]);
 
                 if ($update) {
-                    return response()->json([
-                        'message' => 'Data successfully changed',
-                    ]);
+                    return $this->successResponse(null,'updated');
                 }
+
             }else{
                 $update = $data->update([
                     ...$request->all(),
                     'jabatan_id' => (int) $request->jabatan_id,
                 ]);
                 if ($update) {
-                    return response()->json([
-                        'message' => 'Data successfully changed',
-                    ]);
+                    return $this->successResponse(null,'updated');
                 }
             }
         }
     }
 
-    /* FUNGSI HAPUS ANGGOTA */
     public function destroy($anggota)
     {
         $data = Anggota::findOrfail($anggota);
         if ($data) {
-            Storage::delete($data->foto_anggota);
+            if ($data->foto_anggota != NULL) {
+                Storage::delete($data->foto_anggota);
+            }
             $data->delete();
-            return response()->json([
-                'data' => $data,
-                'message'   => 'Data Berhasil dihapus',
-            ]);
+            return $this->successResponse(null,'deleted');
         }
     }
 }
